+++ zubarev.arsenii/S1/functions.hpp
@@ -9,9 +9,9 @@ namespace zubarev
   void output(std::ostream& out, T el, bool is_first);
 
   size_t max_sequences(List< Data >* list);
-  int output_names(List< Data >* list);
-  int output_sequences(List< Data >* list);
-  int output_sums(List< Data >* list);
+  size_t output_names(List< Data >* list);
+  size_t output_sequences(List< Data >* list);
+  size_t output_sums(List< Data >* list);
   size_t to_size_t(const std::string& num);
   bool isdigit(const std::string& str);
